+++ lib/rails_bootstrap_form/inputs/collection_select.rb
@@ -9,10 +9,10 @@ module RailsBootstrapForm
 
       included do
         def collection_select(attribute, collection, value_method, text_method, options = {}, html_options = {})
-          options = {bootstrap: {field_class: "form-select"}}.deep_merge!(options)
-
           bootstrap = bootstrap_form_options.scoped(options.delete(:bootstrap))
           return super if bootstrap.disabled?
+
+          bootstrap.set_field_class!("form-select")
 
           field_wrapper_builder(attribute, bootstrap, options, html_options) do
             super(attribute, collection, value_method, text_method, options, html_options)